+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ This document records the main changes to the sdss_access code.
 - Breaking Change: switches `sdss_access` to use new SDSS dtn.sdss.org server
 - Feature: adds new "data.sdss5.org" server to support rsync, curl, and http downloads for sdss5 products
 - Adding new Sphinx docs for SDSS collaboration authentication
+- Updates the public release check to now check todays date against the tree release date, and "DR" in release name. 
 
 1.1.1 (2020-11-11)
 ------------------